+++ crypto/CMakeLists.txt
@@ -368,7 +368,6 @@ add_library(
   base64/base64.c
   bio/bio.c
   bio/bio_mem.c
-  bio/cipher.c
   bio/connect.c
   bio/errno.c
   bio/fd.c
@@ -461,6 +460,7 @@ add_library(
   pem/pem_pkey.c
   pem/pem_x509.c
   pem/pem_xaux.c
+  pkcs7/bio/cipher.c
   pkcs7/pkcs7.c
   pkcs7/pkcs7_asn1.c
   pkcs7/pkcs7_x509.c
@@ -759,7 +759,6 @@ if(BUILD_TESTING)
     abi_self_test.cc
     asn1/asn1_test.cc
     base64/base64_test.cc
-    bio/bio_deprecated_test.cc
     bio/bio_test.cc
     blake2/blake2_test.cc
     buf/buf_test.cc
@@ -816,6 +815,7 @@ if(BUILD_TESTING)
     obj/obj_test.cc
     ocsp/ocsp_test.cc
     pem/pem_test.cc
+    pkcs7/bio/bio_deprecated_test.cc
     pkcs7/pkcs7_test.cc
     pkcs8/pkcs8_test.cc
     pkcs8/pkcs12_test.cc
