+++ crypto/bio/internal.h
@@ -114,16 +114,6 @@ int bio_socket_should_retry(int return_value);
 int bio_errno_should_retry(int return_value);
 
 
-// BIO_f_cipher is used internally by the pkcs7 module. It is not recommended
-// for external use.
-OPENSSL_EXPORT const BIO_METHOD *BIO_f_cipher(void);
-
-// BIO_set_cipher is used internally for testing. It is not recommended for
-// external use.
-OPENSSL_EXPORT int BIO_set_cipher(BIO *b, const EVP_CIPHER *cipher,
-                     const unsigned char *key, const unsigned char *iv, int enc);
-
-
 #if defined(__cplusplus)
 }  // extern C
 #endif
